requirements.txt
Django
requests         
python-dotenv    
boto3            
